requirements/development.txt
-r base.txt
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.1
Faker==26.0.0
django-debug-toolbar==4.4.2
